Began work on enemy encounters
Added a battle canvas that is called when the player gets too close to an enemy
Added a lot of AI for enemy movemenet. Attempts to be random and sporadic with no patterns. Plans for fully fleshing out some oddities, but for now the system is great.
Enemy script works with any enemy, just needs a name and battleSprite. Will add new sprites soon.
If the player gets too close to an enemy, the enemy will charge the player. When the enemy is within a set distance of the player, the battle canvas will initiate and battle begins. There are no pauses to action on the map or controls as of yet.
Added background images for battle canvas.